backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
openai
google-cloud-storage
pydantic
python-multipart

